renamed all packages to 'command'
