+++ core/transfer/src/main/java/org/eclipse/dataspaceconnector/transfer/core/command/CommandHandlerRegistryImpl.java
@@ -11,7 +11,7 @@
  *       Microsoft Corporation - initial API and implementation
  *
  */
-package org.eclipse.dataspaceconnector.transfer.core.commandqueue;
+package org.eclipse.dataspaceconnector.transfer.core.command;
 
 import org.eclipse.dataspaceconnector.spi.command.Command;
 import org.eclipse.dataspaceconnector.spi.command.CommandHandler;
