+++ core/transfer/src/main/java/org/eclipse/dataspaceconnector/transfer/core/command/handlers/CancelTransferCommandHandler.java
@@ -11,11 +11,11 @@
  *       Microsoft Corporation - initial API and implementation
  *
  */
-package org.eclipse.dataspaceconnector.transfer.core.commandqueue.handlers;
+package org.eclipse.dataspaceconnector.transfer.core.command.handlers;
 
 import org.eclipse.dataspaceconnector.spi.transfer.store.TransferProcessStore;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcess;
-import org.eclipse.dataspaceconnector.transfer.core.commandqueue.commands.CancelTransferCommand;
+import org.eclipse.dataspaceconnector.transfer.core.command.commands.CancelTransferCommand;
 
 public class CancelTransferCommandHandler extends TransferProcessCommandHandler<CancelTransferCommand> {
