+++ core/transfer/src/main/java/org/eclipse/dataspaceconnector/transfer/core/command/handlers/DeprovisionRequestHandler.java
@@ -1,8 +1,8 @@
-package org.eclipse.dataspaceconnector.transfer.core.commandqueue.handlers;
+package org.eclipse.dataspaceconnector.transfer.core.command.handlers;
 
 import org.eclipse.dataspaceconnector.spi.transfer.store.TransferProcessStore;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcess;
-import org.eclipse.dataspaceconnector.transfer.core.commandqueue.commands.DeprovisionRequest;
+import org.eclipse.dataspaceconnector.transfer.core.command.commands.DeprovisionRequest;
 
 public class DeprovisionRequestHandler extends TransferProcessCommandHandler<DeprovisionRequest> {
