+++ core/transfer/src/test/java/org/eclipse/dataspaceconnector/transfer/core/command/handlers/DeprovisionCommandHandlerTest.java
@@ -1,3 +1,4 @@
+
 /*
  *  Copyright (c) 2020-2022 Microsoft Corporation
  *
@@ -11,13 +12,13 @@
  *       Microsoft Corporation - initial API and implementation
  *
  */
-package org.eclipse.dataspaceconnector.transfer.core.commandqueue.handlers;
+package org.eclipse.dataspaceconnector.transfer.core.command.handlers;
 
 import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.transfer.store.TransferProcessStore;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcess;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcessStates;
-import org.eclipse.dataspaceconnector.transfer.core.commandqueue.commands.CancelTransferCommand;
+import org.eclipse.dataspaceconnector.transfer.core.command.commands.DeprovisionRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -27,40 +28,50 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-class CancelTransferCommandHandlerTest {
+class DeprovisionCommandHandlerTest {
 
-    private CancelTransferCommandHandler handler;
     private TransferProcessStore storeMock;
+    private DeprovisionRequestHandler handler;
 
     @BeforeEach
     void setUp() {
         storeMock = mock(TransferProcessStore.class);
-        handler = new CancelTransferCommandHandler(storeMock);
+        handler = new DeprovisionRequestHandler(storeMock);
     }
 
     @Test
     void handle() {
-        var cmd = new CancelTransferCommand("test-id");
-        var tp = TransferProcess.Builder.newInstance().id("test-id").state(TransferProcessStates.IN_PROGRESS.code())
+        var cmd = new DeprovisionRequest("test-id");
+        var tp = TransferProcess.Builder.newInstance().id("test-id").state(TransferProcessStates.COMPLETED.code())
                 .type(TransferProcess.Type.CONSUMER).build();
 
         when(storeMock.find(anyString())).thenReturn(tp);
         handler.handle(cmd);
 
-        assertThat(tp.getState()).isEqualTo(TransferProcessStates.ERROR.code());
-        assertThat(tp.getErrorDetail()).isEqualTo("Cancelled");
+        assertThat(tp.getState()).isEqualTo(TransferProcessStates.DEPROVISIONING_REQ.code());
+        assertThat(tp.getErrorDetail()).isNull();
     }
 
     @Test
     void handle_notFound() {
-        var cmd = new CancelTransferCommand("test-id");
+        var cmd = new DeprovisionRequest("test-id");
 
         when(storeMock.find(anyString())).thenReturn(null);
-        assertThatThrownBy(() -> handler.handle(cmd)).isInstanceOf(EdcException.class).hasMessageStartingWith("Could not find TransferProcess with ID [test-id]");
+        assertThatThrownBy(() -> handler.handle(cmd)).isInstanceOf(EdcException.class).hasMessage("Could not find TransferProcess with ID [test-id]");
     }
 
     @Test
-    void verifyCorrectType() {
-        assertThat(handler.getType()).isEqualTo(CancelTransferCommand.class);
+    void handle_transitionNotAllowed() {
+        var cmd = new DeprovisionRequest("test-id");
+        var tp = TransferProcess.Builder.newInstance().id("test-id").state(TransferProcessStates.IN_PROGRESS.code())
+                .type(TransferProcess.Type.CONSUMER).build();
+
+        when(storeMock.find(anyString())).thenReturn(tp);
+        assertThatThrownBy(() -> handler.handle(cmd)).isInstanceOf(IllegalStateException.class).hasMessage("Cannot transition from state IN_PROGRESS to DEPROVISIONING_REQ");
+    }
+
+    @Test
+    void getType() {
+        assertThat(handler.getType()).isEqualTo(DeprovisionRequest.class);
     }
 }
